+++ lizard_security/manager.py
@@ -6,6 +6,8 @@ object manager: ``FilteredManager``. We have to set that object manager on our
 models.
 
 """
+from django.contrib.gis.db.models import GeoManager
+from django.db.models.manager import Manager
 from django.db.models.manager import Manager
 from django.db.models import Q
 from tls import request
@@ -25,7 +27,7 @@ def data_set_filter(model_class):
         user = request.user
     except RuntimeError:
         # We don't have a local request object.
-        return empty_data_set
+        return
     if user is not None and user.is_superuser:
         return
     data_set_ids = getattr(request, ALLOWED_DATA_SET_IDS, None)
@@ -44,6 +46,21 @@ class FilteredManager(Manager):
         """Return base queryset, filtered through lizard-security's mechanism.
         """
         query_set = super(FilteredManager, self).get_query_set()
+        extra_filter = data_set_filter(self.model)
+        if extra_filter is not None:
+            query_set = query_set.filter(extra_filter)
+        return query_set
+
+
+class FilteredGeoManager(GeoManager):
+    """Custom geomanager that filters out objects whose data set we can't
+    access.
+    """
+
+    def get_query_set(self):
+        """Return base queryset, filtered through lizard-security's mechanism.
+        """
+        query_set = super(FilteredGeoManager, self).get_query_set()
         extra_filter = data_set_filter(self.model)
         if extra_filter is not None:
             query_set = query_set.filter(extra_filter)